requirements.txt
# Marker Master System - Requirements
# Semantisch-psychologischer Resonanz- und Manipulations-Detektor

# Core dependencies
PyYAML>=6.0
colorama>=0.4.6

# API dependencies (optional)
flask>=2.3.0
flask-cors>=4.0.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional advanced features
# numpy>=1.24.0  # Für erweiterte Statistiken
# scikit-learn>=1.3.0  # Für ML-basierte Erkennung
# spacy>=3.6.0  # Für NLP-Analysen 